refactor(release-manager): BEC-196 — split scheduler.ts (566 lines)
Splits release-manager/scheduler.ts into:
- release-tick.ts — per-tick orchestration (the loop body)
- release-helpers.ts — extracted persistDecision, consumeApprovalRow,
  maybePostSlack, tryFileQaGapIssue, getMaxAttemptCountForReason
- scheduler.ts — thin wrapper that constructs TickMutableState and
  calls the new tick fn each interval

No behavior changes except one intentional improvement:

**Bounded eviction for auditedCompletedRunIds.** The pre-split scheduler
held the set unbounded — fine in test, leaks memory in long-running
deployments. The split was the right moment to fix it because state is
now passed explicitly via TickMutableState. After eviction a single
QA run id can be re-audited once (duplicate qa.run_completed event);
acceptable in exchange for bounded memory. The 10k cap is high enough
to require months of continuous QA runs. Inline comment notes this.

Other gaps from code review:
- Removed stale `release-manager/scheduler.ts` entry from the
  `logAuditEventUnchecked` allowlist in audit-immutability.test.ts
  (the file no longer calls that function). Added the new
  release-tick.ts and release-helpers.ts files in its place.

Tests: 80/80 release-manager tests pass; audit-immutability still green.

Replaces #267 (original branch was based on ee3895e — behind current main
and entangled with unrelated stop/halt files from #268).
